backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
